Added today button
