+++ frontend/src/Component/History/QuickFilter.jsx
@@ -1,0 +1,21 @@
+import React, { useState, useEffect } from "react";
+
+export const QuickFilter = ({ onFilterChange }) => {
+    const [beforeDate, setBeforeDate] = useState("");
+    const [afterDate, setAfterDate] = useState("");
+
+    const Today = () => {
+        let today = new Date().toJSON().slice(0, 10);
+        setBeforeDate(today);
+        setAfterDate(today);
+        onFilterChange({ beforeDate, afterDate });
+    }
+
+    return (
+        <div>
+            <button onClick={Today} type="button" class="py-2.5 px-5 me-2 mb-2 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700">Today</button>
+            <button type="button" class="py-2.5 px-5 me-2 mb-2 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700">This week</button>
+            <button type="button" class="py-2.5 px-5 me-2 mb-2 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700">This month</button>
+        </div>
+    );
+};
